+++ Cargo.toml
@@ -41,6 +41,7 @@ members = [
     "src/storage",
     "src/storage/backup",
     "src/storage/compactor",
+    "src/storage/hummock_memory_limiter",
     "src/storage/hummock_sdk",
     "src/storage/hummock_test",
     "src/stream",
@@ -214,6 +215,7 @@ risingwave_error = { path = "./src/error" }
 risingwave_expr = { path = "./src/expr/core" }
 risingwave_expr_impl = { path = "./src/expr/impl" }
 risingwave_frontend = { path = "./src/frontend" }
+risingwave_hummock_memory_limiter = { path = "./src/storage/hummock_memory_limiter" }
 risingwave_hummock_sdk = { path = "./src/storage/hummock_sdk" }
 risingwave_hummock_test = { path = "./src/storage/hummock_test" }
 risingwave_hummock_trace = { path = "./src/storage/hummock_trace" }
