+++ src/storage/Cargo.toml
@@ -47,6 +47,7 @@ risingwave_common = { workspace = true }
 risingwave_common_estimate_size = { workspace = true }
 risingwave_common_service = { workspace = true }
 risingwave_expr = { workspace = true }
+risingwave_hummock_memory_limiter = { workspace = true }
 risingwave_hummock_sdk = { workspace = true }
 risingwave_hummock_trace = { workspace = true }
 risingwave_object_store = { workspace = true }
