+++ src/storage/hummock_memory_limiter/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "risingwave_hummock_memory_limiter"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["workspace-hack", "num-traits"]
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["workspace-hack"]
+
+[dev-dependencies]
+rand = { workspace = true }
+futures = { version = "0.3" }
+
+[target.'cfg(madsim)'.dependencies]
+tokio = { version = "0.2", package = "madsim-tokio", features = ["sync"]}
+
+[target.'cfg(not(madsim))'.dependencies]
+tokio = {version = "^1.36.0", features = ["sync", "rt", "rt-multi-thread", "time", "macros"]}
+workspace-hack = { path = "../../workspace-hack" }
+
+[target."cfg(not(loom))".dependencies]
+parking_lot = { version = "0.12.3", features = ["deadlock_detection"] }
+
+[target."cfg(loom)".dependencies]
+loom = { version = "0.7" , features = ["futures", "checkpoint"]}
+
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
